feat(neovim): set Tree-sitter/Query specific options
Set window local options using `winid` and `0` to ensure options are
local, avoiding leaking this option values into other buffers.

Enable spellchecking, except for in unmodifiable and `nofile` buffers as
those are probably not meant to be edited anyway. Also disable number
and sign line in `:InspectTree` to leave more room for the tree.