docs(notes)!: merge most notes into install and tools
This removes a lot of clutter that has been accumulating in the notes
directory.
Now most things should belong to install or tools. The tmux notes are
kept around since they might come in handy as quick refresh if I have to
use tmux any time again.
Also took the time to update and/or remove outdated statements.